Gemfile for Puppet/Facter and ignore Gemfile.lock
This will, by default, install the latest minor version of Puppet 3.1.x and
Facter 1.6.x for local development purposes.

These versions can be overridden with environment variables so that tests
can be run against a matrix of different Puppet, Facter, and Ruby versions.
In order to do this, `Gemfile.lock` which pins specific versions, has been
omitted and added to `.gitignore`. The major version pinning is done using
Bundler's `~>` notation.

We might want to consider upgrading to Facter 1.7, however it is probably
not worth evaluating Puppet 3.2 just yet.