Travis CI tests for Puppet and Ruby versions
Add matrix tests for our targetted version (Puppet 3.1 on Ruby 1.9.3) and
future versions (Puppet 3.2 on Ruby HEAD). The latter are allowed to fail,
as an indicator of what may be to come.

Omitting older versions of Puppet and Ruby for the moment. We could also
add future versions of Facter in here.

The default rake task will be run. Which currently only contains lint.