Move third-party modules to vendor/modules
Ensures that they are kept separate. First party modules can be committed to
this repo, on the basis that they are too small, trivial, or organisation
specific, to break out with librarian.

Destructive is set to true. Since this path shouldn't contain any content
that isn't managed directly by librarian-puppet.

/cc @nickstenning @samjsharpe