Adjust RAM for Vagrant nodes
The new PuppetLabs base box introduced in ab1fd8a has a default of 2048M
RAM. That's quite thirsty if you plan to run three or more nodes at the same
time.

Bump them all down to a default of 384M and specify one as 512M to
demonstrate that it can be overridden.