Deprecate `import` statement
Puppet announced that `import` is being fully deprecated by Puppet 4's
released date. Currently, we make use of this in this repository to import
the content of `nodes.pp` (which then looks up the values associated with
the `classes` key in Hiera).

The deprecation notice is available at
http://docs.puppetlabs.com/puppet/latest/reference/lang_import.html#deprecation-notice.
This commit allows us to follow the first of the two suggested workarounds.