Upgrade to Puppet Forge API version 3
Change the URL in `Puppetfile` and `Puppetfile.lock` so that we're using
v3 of the Puppet Forge API.

This also has the benefit of providing transport layer security (HTTPS)
for the API connection, which gives us some guarantee that we are
connecting to the correct server.

The line `LIBRARIAN_PUPPET_USE_V1_API: '1'` was added when I ran
`bundle exec librarian-puppet install`. I can set it to false by running:

```sh
bundle exec librarian-puppet install --no-use-v1-api
```

`--use-v1-api` is the default[1][2] but libarian-puppet ignores this
setting for the official Puppet Forge[3] and uses the v3 API. So leave
it as-is to make running `librarian-puppet install` simpler.

[1]: https://github.com/rodjek/librarian-puppet/tree/2f0423e883b374ff805500acf1ab9986f2690e1d#how-to-use
[2]: https://github.com/rodjek/librarian-puppet/blob/2f0423e883b374ff805500acf1ab9986f2690e1d/lib/librarian/puppet/cli.rb#L64
[3]: https://github.com/rodjek/librarian-puppet/blob/2f0423e883b374ff805500acf1ab9986f2690e1d/lib/librarian/puppet/source/forge.rb#L154