bump puppetlabs-stdlib to 4.5.1 because of CVE-2015-1029
Versions of puppetlabs-stdlib prior to 4.5.1 suffered from a
vulnerability [1] where a nonprivileged user could pre-populate the
stdlib fact cache, potentially resulting in changing the behaviour of
the puppet run.

I bumped the Puppetfile too to exclude as many vulnerable versions as
possible without preventing upgrade to 4.6.  4.5.0 is still
allowed but that's the limitation of the ~> operator, and with 4.5.1 in
Puppetfile.lock it's unlikely a project started from this skeleton will
ever use 4.5.0.

[1] http://puppetlabs.com/security/cve/cve-2015-1029